+++ code/sspp-assignment/common/CRSTransformer.h
@@ -17,28 +17,28 @@ namespace sspp {
       static CRS<VALUE_TYPE> transform(MatrixMarket<INPUT_VALUE> & matrix_market) {
         std::vector<MatrixMarketTuple<INPUT_VALUE>> tuples = matrix_market.GetTuples();
         std::sort(tuples.begin(), tuples.end(), MatrixMarket<INPUT_VALUE>::CompareMatrixMarketTuple);
-        std::vector<unsigned> column_indices(matrix_market.GetNonZeros()), row_start_indexes(matrix_market.GetRows() + 1);
+        std::vector<unsigned long long> column_indices(matrix_market.GetNonZeros()), row_start_indexes(matrix_market.GetRows() + 1);
         std::vector<VALUE_TYPE> values(matrix_market.GetNonZeros());
 
-        unsigned row_start_indexes_index = 0, non_zeros_index = 1;
+        unsigned long long row_start_indexes_index = 0, non_zeros_index = 1;
         row_start_indexes[row_start_indexes_index++] = 0;
 
         while(row_start_indexes_index < matrix_market.GetRows() + 1
               && non_zeros_index < matrix_market.GetNonZeros()) {
           column_indices[non_zeros_index - 1] = tuples[non_zeros_index - 1].GetColumnIndice();
           values[non_zeros_index - 1] = static_cast<VALUE_TYPE>(tuples[non_zeros_index - 1].GetValue());
-          unsigned row_indices_diff = tuples[non_zeros_index].GetRowIndice() - tuples[non_zeros_index - 1].GetRowIndice();
+          unsigned long long row_indices_diff = tuples[non_zeros_index].GetRowIndice() - tuples[non_zeros_index - 1].GetRowIndice();
           if(row_indices_diff != 0) {
             row_start_indexes[row_start_indexes_index++] = non_zeros_index;
             if(row_indices_diff > 1) {
-              for(unsigned i = 0; i < row_indices_diff - 1; ++i) {
+              for(unsigned long long i = 0; i < row_indices_diff - 1; ++i) {
                 row_start_indexes[row_start_indexes_index++] = non_zeros_index;
               }
             }
           }
           ++non_zeros_index;
         }
-        for(unsigned i = row_start_indexes_index; i < matrix_market.GetRows(); ++i) {
+        for(unsigned long long i = row_start_indexes_index; i < matrix_market.GetRows(); ++i) {
           row_start_indexes[row_start_indexes_index++] = non_zeros_index;
         }
         row_start_indexes[row_start_indexes_index++] = non_zeros_index;
