+++ code/sspp-assignment/commonTests/AdvancedCRSTests.cpp
@@ -12,8 +12,8 @@ using namespace sspp::common;
 
 TEST_F(CSRTransformerTest, ALL_ROWS) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 3, N = 3, NZ = 5;
-  std::vector<unsigned> iIndexes = { 0, 0, 1, 1, 2 },
+  const unsigned long long M = 3, N = 3, NZ = 5;
+  std::vector<unsigned long long> iIndexes = { 0, 0, 1, 1, 2 },
     jIndexes = { 0, 1, 1, 2, 2 };
   std::vector<float> values = { 17, 4, 5, 2, 11 };
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
@@ -33,8 +33,8 @@ TEST_F(CSRTransformerTest, ALL_ROWS) {
 
 TEST_F(CSRTransformerTest, NO_MIDDLE_ROW) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 3, N = 3, NZ = 3;
-  std::vector<unsigned> iIndexes = { 0, 0, 2 },
+  const unsigned long long M = 3, N = 3, NZ = 3;
+  std::vector<unsigned long long> iIndexes = { 0, 0, 2 },
     jIndexes = { 0, 1, 2 };
   std::vector<float> values = { 17, 4, 11 };
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
@@ -54,8 +54,8 @@ TEST_F(CSRTransformerTest, NO_MIDDLE_ROW) {
 
 TEST_F(CSRTransformerTest, NO_MIDDLE_ROW_2) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 3, N = 3, NZ = 2;
-  std::vector<unsigned> iIndexes = { 0, 2 },
+  const unsigned long long M = 3, N = 3, NZ = 2;
+  std::vector<unsigned long long> iIndexes = { 0, 2 },
     jIndexes = { 1, 2 };
   std::vector<float> values = { 4, 11 };
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
@@ -75,8 +75,8 @@ TEST_F(CSRTransformerTest, NO_MIDDLE_ROW_2) {
 
 TEST_F(CSRTransformerTest, ONE_VALUE_ONLY) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 3, N = 3, NZ = 1;
-  std::vector<unsigned> iIndexes = { 2 },
+  const unsigned long long M = 3, N = 3, NZ = 1;
+  std::vector<unsigned long long> iIndexes = { 2 },
     jIndexes = { 2 };
   std::vector<float> values = { 11 };
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
@@ -96,8 +96,8 @@ TEST_F(CSRTransformerTest, ONE_VALUE_ONLY) {
 
 TEST_F(CSRTransformerTest, FIRST_AND_LAST_ROW) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 4, N = 4, NZ = 2;
-  std::vector<unsigned> iIndexes = { 0, 3 },
+  const unsigned long long M = 4, N = 4, NZ = 2;
+  std::vector<unsigned long long> iIndexes = { 0, 3 },
     jIndexes = { 0, 3 };
   std::vector<float> values = { 4, 11};
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
@@ -117,8 +117,8 @@ TEST_F(CSRTransformerTest, FIRST_AND_LAST_ROW) {
 
 TEST_F(CSRTransformerTest, FIRST_AND_LAST_ROW_2) {
   MatrixMarketHeader mmh(Matrix, Sparse, Real, General);
-  const unsigned M = 4, N = 4, NZ = 4;
-  std::vector<unsigned> iIndexes = { 0, 0, 0, 3 },
+  const unsigned long long M = 4, N = 4, NZ = 4;
+  std::vector<unsigned long long> iIndexes = { 0, 0, 0, 3 },
     jIndexes = { 0, 1, 2, 3 };
   std::vector<float> values = { 4, 5, 6, 11 };
   MatrixMarket<float> mm(M, N, NZ, iIndexes, jIndexes, values);
